Moving SearchModule to search module and removing the dependencies that are not needed
